Release Version

+++ backend/src/models/product.py
@@ -1,9 +1,10 @@
 from json import loads
-from backend.src.lib import Global
+
+from backend.src.lib import Global, Result
+from backend.src.lib.db import create_connection
+from backend.src.models.category import Category
 from backend.src.models.model import Model
 from backend.src.models.user import User
-from backend.src.models.category import Category
-from backend.src.lib import Result
 
 
 class Product(Model):
@@ -17,12 +18,19 @@ class Product(Model):
         sort_newest: bool = False,
     ) -> list[dict[str]]:
         taking = list(__taking)
-        db_conn = Global.db_conn
+        __db_conn = create_connection()
+        if __db_conn.is_err():
+            Global.console.print(str(__db_conn.unwrap_err()))
+            return []
+
+        db_conn = __db_conn.unwrap()
+
         cond, args = Product.criteria_to_arguments(__cond)
         if len(taking) == 0:
             sb = f"""SELECT p.id, p.pid, p.name, p.images, c.name as catName,
-u.username as ownerName, p.price, p.customization, p.rating, p.availability, p.soldAmount,
-p.deliveryOption, p.description, p.createdAt FROM products AS p
+u.id as ownerId, u.username as ownerName, p.price, p.customization, p.rating, p.availability, p.soldAmount,
+p.deliveryOption, p.description, p.createdAt, p.deleted
+FROM products AS p
 INNER JOIN categories AS c ON p.catId = c.id
 INNER JOIN users as u ON p.owner = u.id
 WHERE {cond}
@@ -42,8 +50,6 @@ WHERE {cond}
             if limit != -1:
                 sb += " LIMIT %s OFFSET %s"
 
-        print(sb)
-
         cursor = db_conn.cursor(prepared=True)
 
         # * scuffed code, can't be bothered to fix
@@ -60,12 +66,18 @@ WHERE {cond}
     @classmethod
     def fetch_all(cls, __taking: list[str]) -> list[dict[str]]:
         taking = list(__taking)
-        db_conn = Global.db_conn
+        __db_conn = create_connection()
+        if __db_conn.is_err():
+            Global.console.print(str(__db_conn.unwrap_err()))
+            return []
+
+        db_conn = __db_conn.unwrap()
 
         if len(taking) == 0:
             sb = """SELECT p.id, p.pid, p.name, p.images, c.name as catName,
-u.username as ownerName, p.price, p.customization, p.rating, p.availability, p.soldAmount,
-p.deliveryOption, p.description, p.createdAt FROM products AS p
+u.id as ownerId, u.username as ownerName, p.price, p.customization, p.rating, p.availability, p.soldAmount,
+p.deliveryOption, p.description, p.createdAt, p.deleted
+FROM products AS p
 INNER JOIN categories AS c ON p.catId = c.id
 INNER JOIN users as u ON p.owner = u.id"""
         else:
@@ -88,12 +100,18 @@ INNER JOIN users as u ON p.owner = u.id"""
         __taking: list[str], offset: int = 0, limit: int = 25
     ) -> list[dict[str]]:
         taking = list(__taking)
-        db_conn = Global.db_conn
+        __db_conn = create_connection()
+        if __db_conn.is_err():
+            Global.console.print(str(__db_conn.unwrap_err()))
+            return []
+
+        db_conn = __db_conn.unwrap()
 
         if len(taking) == 0:
             sb = """SELECT p.id, p.pid, p.name, p.images, c.name as catName,
-u.username as ownerName, p.price, p.customization, p.rating, p.availability, p.soldAmount,
-p.deliveryOption, p.description, p.createdAt FROM products AS p
+u.id as ownerId, u.username as ownerName, p.price, p.customization, p.rating, p.availability, p.soldAmount,
+p.deliveryOption, p.description, p.createdAt, p.deleted
+FROM products AS p
 INNER JOIN categories AS c ON p.catId = c.id
 INNER JOIN users as u ON p.owner = u.id
 ORDER BY p.createdAt DESC
@@ -114,11 +132,18 @@ LIMIT %s OFFSET %s"""
 
     @classmethod
     def id(cls, __id: str | int) -> dict[str]:
-        db_conn = Global.db_conn
+        __db_conn = create_connection()
+        if __db_conn.is_err():
+            Global.console.print(str(__db_conn.unwrap_err()))
+            return []
+
+        db_conn = __db_conn.unwrap()
+
         cursor = db_conn.cursor(prepared=True)
         sql = """SELECT p.id, p.pid, p.name, p.images, c.name as catName,
-u.username as ownerName, p.price, p.customization, p.rating, p.availability, p.soldAmount,
-p.deliveryOption, p.description, p.createdAt FROM products AS p
+u.id as ownerId, u.username as ownerName, p.price, p.customization, p.rating, p.availability, p.soldAmount,
+p.deliveryOption, p.description, p.createdAt, p.deleted
+FROM products AS p
 INNER JOIN categories AS c ON p.catId = c.id
 INNER JOIN users as u ON p.owner = u.id
 WHERE p.id = %s"""
@@ -133,11 +158,18 @@ WHERE p.id = %s"""
 
     @classmethod
     def pid(cls, __pid: str) -> dict[str]:
-        db_conn = Global.db_conn
+        __db_conn = create_connection()
+        if __db_conn.is_err():
+            Global.console.print(str(__db_conn.unwrap_err()))
+            return []
+
+        db_conn = __db_conn.unwrap()
+
         cursor = db_conn.cursor(prepared=True)
         sql = """SELECT p.id, p.pid, p.name, p.images, c.name as catName,
-u.username as ownerName, p.price, p.customization, p.rating, p.availability, p.soldAmount,
-p.deliveryOption, p.description, p.createdAt FROM products AS p
+u.id as ownerId, u.username as ownerName, p.price, p.customization, p.rating, p.availability, p.soldAmount,
+p.deliveryOption, p.description, p.createdAt, p.deleted
+FROM products AS p
 INNER JOIN categories AS c ON p.catId = c.id
 INNER JOIN users as u ON p.owner = u.id
 WHERE p.pid = %s"""
@@ -162,6 +194,7 @@ WHERE p.pid = %s"""
                 "name",
                 "images",
                 "catName",
+                "ownerId",
                 "ownerName",
                 "price",
                 "customization",
@@ -171,6 +204,7 @@ WHERE p.pid = %s"""
                 "deliveryOption",
                 "description",
                 "createdAt",
+                "deleted",
             ]
 
         res = {taking[i]: __result[i] for i in range(len(taking))}
@@ -194,7 +228,12 @@ WHERE p.pid = %s"""
             sb += ", ".join(["%s" for _ in range(len(keys_keys))])
             sb += ")"
 
-            db_conn = Global.db_conn
+            __db_conn = create_connection()
+            if __db_conn.is_err():
+                Global.console.print(str(__db_conn.unwrap_err()))
+                return []
+
+            db_conn = __db_conn.unwrap()
             cursor = db_conn.cursor(prepared=True)
             cursor.execute(sb, tuple([keys[key] for key in keys_keys]))
             db_conn.commit()
@@ -213,7 +252,13 @@ WHERE p.pid = %s"""
             sb += ", ".join([f"{key} = %s" for key in keys_keys])
             sb += " WHERE pid = %s"
 
-            db_conn = Global.db_conn
+            __db_conn = create_connection()
+            if __db_conn.is_err():
+                Global.console.print(str(__db_conn.unwrap_err()))
+                return []
+
+            db_conn = __db_conn.unwrap()
+
             cursor = db_conn.cursor(prepared=True)
             cursor.execute(sb, tuple([keys[key] for key in keys_keys] + [pid]))
             db_conn.commit()
@@ -222,23 +267,42 @@ WHERE p.pid = %s"""
         except Exception as e:
             Global.console.print_exception()
             return Result.Err(str(e))
-        
+
     @staticmethod
     def delete(pid: str) -> Result[tuple, str]:
         try:
-            db_conn = Global.db_conn
-            cursor = db_conn.cursor(prepared=True)
-            cursor.execute("DELETE FROM products WHERE pid = %s", (pid,))
-            db_conn.commit()
-            cursor.close()
-            return Result.Ok(())
+            __db_conn = create_connection()
+            if __db_conn.is_err():
+                Global.console.print(str(__db_conn.unwrap_err()))
+                return []
+
+            db_conn = __db_conn.unwrap()
+            try:
+                cursor = db_conn.cursor(prepared=True)
+                cursor.execute("DELETE FROM products WHERE pid = %s", (pid,))
+                db_conn.commit()
+                cursor.close()
+                return Result.Ok(())
+            except Exception as e:
+                # try soft delete instead
+                cursor = db_conn.cursor(prepared=True)
+                cursor.execute("UPDATE products SET deleted = 1 WHERE pid = %s", (pid,))
+                db_conn.commit()
+                return Result.Ok(())
+
         except Exception as e:
             Global.console.print_exception()
             return Result.Err(str(e))
-    
+
     @staticmethod
     def attest_nonexistent(pid: str) -> bool:
-        db_conn = Global.db_conn
+        __db_conn = create_connection()
+        if __db_conn.is_err():
+            Global.console.print(str(__db_conn.unwrap_err()))
+            return []
+
+        db_conn = __db_conn.unwrap()
+
         cursor = db_conn.cursor(prepared=True)
         cursor.execute("SELECT COUNT(*) FROM products WHERE pid = %s", (pid,))
         result = cursor.fetchone()
@@ -249,13 +313,41 @@ WHERE p.pid = %s"""
     @staticmethod
     def attest_reviewed(pid: str, uid: str | int) -> bool:
         uid = int(uid)
-        db_conn = Global.db_conn
+        __db_conn = create_connection()
+        if __db_conn.is_err():
+            Global.console.print(str(__db_conn.unwrap_err()))
+            return []
+
+        db_conn = __db_conn.unwrap()
         cursor = db_conn.cursor(prepared=True)
-        cursor.execute("SELECT COUNT(*) FROM reviews WHERE pid = %s AND authorId = %s", (pid, uid))
+        cursor.execute(
+            """\
+SELECT COUNT(*)
+FROM reviews as r
+INNER JOIN products as p ON p.id = r.productId
+WHERE p.pid = %s AND r.authorId = %s""",
+            (pid, uid),
+        )
         result = cursor.fetchone()
         cursor.close()
-        
+
         return result[0] != 0
+
+    @staticmethod
+    def owner(pid: str) -> int:
+        __db_conn = create_connection()
+        if __db_conn.is_err():
+            Global.console.print(str(__db_conn.unwrap_err()))
+            return []
+
+        db_conn = __db_conn.unwrap()
+
+        cursor = db_conn.cursor(prepared=True)
+        cursor.execute("SELECT owner FROM products WHERE pid = %s", (pid,))
+        result = cursor.fetchone()
+        cursor.close()
+
+        return result[0]
 
     @staticmethod
     def __canonical_name(k: str) -> str:
@@ -263,6 +355,8 @@ WHERE p.pid = %s"""
             return "c.name"
         elif k == "ownerName":
             return "u.username"
+        elif k == "ownerId":
+            return "u.id"
         else:
             return f"p.{k}"
